feat(api): add sendCommand method to Api class
Adds a new method, sendCommand, to the Api class. This method is used to send a command to the server. The is passed as a string

The method first logs the command being sent using the DEBUG_INFO macro. Then, it if the command was successfully sent to the server using the protocol_client_send_packet function. If the command fails to send, an ApiException is thrown.

Finally, the method logs a